AI/VisionAgents/Lyria3/pyproject.toml
[project]
name = "lyria3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "vision-agents[deepgram,elevenlabs,gemini,getstream,openai]>=0.3.7",
    "google-genai>=1.16.0",
    "vision-agents-plugins-gemini>=0.3.7",
]
